refactor event tests; finish writing contact list tests
