+++ NosTests/Fixtures/contact_list.json
@@ -7,16 +7,10 @@
             "d0a1ffb8761b974cec4a3be8cbcb2e96a7090dcf465ffeac839aa4ca20c9a59e",
             "wss://nostr.lorentz.is/",
             "Test Name"
-        ],
-        [
-            "p",
-            "76c71aae3a491f1d9eec47cba17e229cda4113a0bbb6e6ae1776d7643e29cafa",
-            "wss://nostr.lorentz.is/",
-            "Another Test Name"
         ]
     ],
-    "created_at": 1,
+    "created_at": 1675264762,
     "pubkey": "32730e9dfcab797caf8380d096e548d9ef98f3af3000542f9271a91a9e3b0001",
-    "id": "abcd1234",
+    "id": "404c87ca6095233cb82b35e2d39d82514124f933addd5514d3dc862b30e4cb4d",
     "sig": "a01fa191a0236ffe5ee1fbd9401cd7b1da7daad5e19a25962eb7ea4c9335522478bdff255f1de40ca6c98cdf8cf26aa1f5f1b6c263c5004b0b6dcdc12573cfd7"
 }
